requirements.txt
# iMessage
imessage-reader

# Gmail
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth

# Supabase
supabase
python-dotenv

# Django
Django
djangorestframework
django-filter

# AWS Comprehend
boto3

# Date parsing
python-dateutil